# Implement NAT port pool manager to prevent src port collisions