--- a/src/org/freertr/user/userClear.java
+++ b/src/org/freertr/user/userClear.java
@@ -56,6 +56,7 @@ import org.freertr.rtr.rtrRip6neigh;
 import org.freertr.serv.servBmp2mrt;
 import org.freertr.serv.servP4lang;
 import org.freertr.tab.tabRouteAttr;
+import org.freertr.tab.tabNatTraN;
 import org.freertr.util.bits;
 import org.freertr.util.cmds;
 import org.freertr.enc.encXml;
@@ -824,6 +825,17 @@ public class userClear {
                     cmd.error("no such vrf");
                     return null;
                 }
+                // Release resources (including port allocations)
+                for (int i = vrf.fwd4.natTrns.size() - 1; i >= 0; i--) {
+                    tabNatTraN ntry = vrf.fwd4.natTrns.get(i);
+                    if (ntry != null) {
+                        ntry.releaseResources();
+                        
+                        if (ntry.logEnd) {
+                            logger.info("removing translation " + ntry);
+                        }
+                    }
+                }
                 vrf.fwd4.natTrns.clear();
                 return null;
             }
@@ -950,6 +962,17 @@ public class userClear {
                 if (vrf == null) {
                     cmd.error("no such vrf");
                     return null;
+                }
+                // Release resources (including port allocations)
+                for (int i = vrf.fwd6.natTrns.size() - 1; i >= 0; i--) {
+                    tabNatTraN ntry = vrf.fwd6.natTrns.get(i);
+                    if (ntry != null) {
+                        ntry.releaseResources();
+                        
+                        if (ntry.logEnd) {
+                            logger.info("removing translation " + ntry);
+                        }
+                    }
                 }
                 vrf.fwd6.natTrns.clear();
                 return null;