Added tabNatPortPoolManager.java to centrally manage NAT src port resources, ensuring unique src port allocation across translations. This prevents potential port collisions that could lead to connection issues.

The implementation includes:
- A singleton class to manage global port allocations
- Support for multiple src port allocation strategies:
  * RandomPortAllocation: Uses random port selection with collision avoidance (ipv4 nat < vrf > sequence < sequence > randomize 16384 32000)
  * SequentialPortAllocation: Allocates ports sequentially from the available range (ipv4 nat < vrf > sequence < sequence > randomize sequential-port-allocation 16384 32000)
  * PreserveOriginalThenSequential: Tries to keep original port, falls back to sequential (ipv4 nat < vrf > sequence < sequence > randomize preserve-original-then-sequential 16384 32000)
  * PreserveOriginalThenRandom: Tries to keep original port, falls back to random (ipv4 nat < vrf > sequence < sequence > randomize preserve-original-then-random 16384 32000)
- Default strategy if not explicitly configured: PreserveOriginalThenRandom
- Port range management per IP address
- Thread-safe allocation and release mechanisms

Enhanced CLI commands:
- show ipv4 nat < vrf > translations: Added sequence numbers and initiator/responder type columns
- show ipv4 nat < vrf > port-pool-usage: New command to display port pool utilization
- show ipv4 nat < vrf > hardware-counters: New command to display hardware counters

Modified related NAT components:
- Updated tabNatCfgN to use port pools for translations
- Enhanced tabNatTraN with better resource tracking and management
- Modified ipFwdTab to handle the new port allocation strategy
- Updated userShow to support the new display commands

This change improves NAT stability and ensures that source port collisions are no longer possible, thereby preventing connection failures, It would be great if you could consider accepting this code or even parts of it.

During debugging, I noticed that only the WAN-to-LAN direction is hardware accelerated using P4 (DPDK), while the LAN-to-WAN direction always uses the slow path for every connection, regardless of whether a NAT translation already exists. Is this intentional? You can observe this behavior with the new command:
show ipv4 nat <vrf> hardware-counters.

BR
Takalele