@mc36 2025-04-27
hi,

On 4/27/25 14:42, Takalele wrote:
> Added tabNatPortPoolManager.java to centrally manage NAT src port resources, ensuring unique src port allocation across translations. This prevents potential port collisions that 
> could lead to connection issues.
> 

nice work, well done! :)

> The implementation includes:
> 
>   * A singleton class to manage global port allocations
let me change this to a per vrf per afi instance instead;

vrfs are fully separated, with separate nat tables, no reason to keep the ports unique globally?

[..]


> During debugging, I noticed that only the WAN-to-LAN direction is hardware accelerated using P4 (DPDK), while the LAN-to-WAN direction always uses the slow path for every 
> connection, regardless of whether a NAT translation already exists. Is this intentional? You can observe this behavior with the new command:
> show ipv4 nat hardware-counters.

nope i cannot spot the behavior nor on my homenet nor in the tests;

cd src
./c.sh
./cn.sh
./twd.sh p4lang-acl004.tst
telnet 127.0.0.1 20001

r1#show ipv4 nat v1 hardware-counters
        src               dest              sw            hw              total
proto  addr       port   addr       port   packs  bytes  packs  bytes    packs  bytes
6      2.2.2.105  4552   2.2.2.101  38987  0      0      115    7592     115    7592
6      2.2.2.106  38987  2.2.2.105  4552   1      48     115    7592     116    7640
17     2.2.2.105  4552   2.2.2.101  4552   0      0      1158   1288663  1158   1288663
17     2.2.2.106  4552   2.2.2.105  4552   1      42     1157   1288607  1158   1288649

r1#


if i reconfigure the test to inbound nat, i still cannot spot the slow-path packets...

br,

cs