@mc36 2025-04-28
hi,

On 4/27/25 16:52, mc36 wrote:

further looking around and i noticed that you're converting the binary represented addresses to strings, so

instead, i came up with this: https://github.com/mc36/freeRtr/blob/master/src/org/freertr/ip/ipFwd.java#L2167

now it seems that it addresses the 'overloading' issue; doing "socat - tcp4:nop.hu:23,bind=:1234" from

multiple inside hosts results in the following translations, seemingly all works just fine:

mchome#show ipv4 nat inet2 translations | include 1234
6      80.211.201.218 23           178.164.182.227 1234        80.211.201.218 23           100.100.5.13 1234           00:06:51  00:00:31  00:05:00  0+11        0+2326
6      80.211.201.218 23           178.164.182.227 42388       80.211.201.218 23           100.100.5.45 1234           00:06:48  00:00:01  00:05:00  0+37        0+8804
6      80.211.201.218 23           178.164.182.227 47660       80.211.201.218 23           100.100.5.5 1234            00:06:49  00:00:16  00:05:00  0+27        0+6882
6      80.211.201.218 23           178.164.182.227 50399       80.211.201.218 23           100.100.5.1 1234            00:06:49  00:00:16  00:05:00  0+36        0+8500
6      80.211.201.218 23           178.164.182.227 53029       80.211.201.218 23           100.100.5.9 1234            00:06:35  00:00:01  00:05:00  0+45        0+9672
6      100.100.5.1 1234            80.211.201.218 23           178.164.182.227 50399       80.211.201.218 23           00:06:49  00:00:16  00:05:00  1+30        41+1674
6      100.100.5.5 1234            80.211.201.218 23           178.164.182.227 47660       80.211.201.218 23           00:06:49  00:00:16  00:05:00  1+22        41+1234
6      100.100.5.9 1234            80.211.201.218 23           178.164.182.227 53029       80.211.201.218 23           00:06:35  00:00:01  00:05:00  1+37        60+2050
6      100.100.5.13 1234           80.211.201.218 23           178.164.182.227 1234        80.211.201.218 23           00:06:51  00:00:31  00:05:00  1+6         41+344
6      100.100.5.45 1234           80.211.201.218 23           178.164.182.227 42388       80.211.201.218 23           00:06:48  00:00:01  00:05:00  1+31        41+1729

mchome#


if you prefer to have other knobs like the sequential port strategy, feel free to add that back...

br,

cs