@Takalele 2025-05-01
Hi @mc36 


I've been testing both NAT implementations and wanted to share my findings with you:

Without Port Pool (https://github.com/mc36/freeRtr/blob/master/src/org/freertr/ip/ipFwd.java#L2167):
- 99.66% success (95 failed out of 28232 connections)
- 161.85s total runtime
- ~984ms average connection time
- Range: 0.41ms to 65.1s
- 99.9th percentile at ~64.8s

With Port Pool:
- 100% success rate
- Much faster at 30.71s runtime
- Better ~684ms average connection time
- Range: 0.76ms to 7s
- 99.9th percentile at ~5.7s

The Port Pool version is about 5x faster, but you know what? 
For my home setup, I'm actually happy with the non-Pool version. The small difference 
in success rate (99.66%) isn't noticeable in normal use, and I prefer the simpler 
implementation. The performance issues only start to show up 
when the port pool utilization reaches 70% or higher, which is unlikely in a home 
setting.

Let me know what you think!

br
Takalele

non-Pool version:
```
[Client] Connection Summary:
  Total connections attempted: 28232
  Successful connections: 28137
  Failed connections: 95
  Success rate: 99.66%
  Total runtime: 161.85 seconds
  Average connection time: 984.20ms
  Fastest connection time: 0.41ms
  Slowest connection time: 65116.64ms

[Client] Per-Decile Statistics:
  Decile 1 (10%):
    Connections: 2813
    Fastest time: 0.41ms
    Slowest time: 4.84ms
    Avg time: 3.36ms
  Decile 2 (20%):
    Connections: 2813
    Fastest time: 4.84ms
    Slowest time: 7.76ms
    Avg time: 6.18ms
  Decile 3 (30%):
    Connections: 2813
    Fastest time: 7.76ms
    Slowest time: 13.79ms
    Avg time: 10.19ms
  Decile 4 (40%):
    Connections: 2813
    Fastest time: 13.80ms
    Slowest time: 168.63ms
    Avg time: 82.45ms
  Decile 5 (50%):
    Connections: 2813
    Fastest time: 168.72ms
    Slowest time: 418.31ms
    Avg time: 271.41ms
  Decile 6 (60%):
    Connections: 2813
    Fastest time: 418.32ms
    Slowest time: 926.21ms
    Avg time: 646.37ms
  Decile 7 (70%):
    Connections: 2813
    Fastest time: 926.26ms
    Slowest time: 1201.26ms
    Avg time: 1050.97ms
  Decile 8 (80%):
    Connections: 2813
    Fastest time: 1201.27ms
    Slowest time: 1885.05ms
    Avg time: 1547.15ms
  Decile 9 (90%):
    Connections: 2813
    Fastest time: 1885.19ms
    Slowest time: 2185.32ms
    Avg time: 2072.67ms
  Decile 10 (100%):
    Connections: 2820
    Fastest time: 2185.35ms
    Slowest time: 65116.64ms
    Avg time: 4143.39ms

[Client] Upper Percentile Statistics:
  90th percentile:
    Connections: 2814
    Fastest time: 2185.99ms
    Slowest time: 65116.64ms
    Avg time: 4147.57ms
  91th percentile:
    Connections: 2533
    Fastest time: 2196.34ms
    Slowest time: 65116.64ms
    Avg time: 4364.56ms
  92th percentile:
    Connections: 2251
    Fastest time: 2205.23ms
    Slowest time: 65116.64ms
    Avg time: 4635.61ms
  93th percentile:
    Connections: 1970
    Fastest time: 2212.18ms
    Slowest time: 65116.64ms
    Avg time: 4981.86ms
  94th percentile:
    Connections: 1689
    Fastest time: 2230.88ms
    Slowest time: 65116.64ms
    Avg time: 5441.60ms
  95th percentile:
    Connections: 1407
    Fastest time: 2261.19ms
    Slowest time: 65116.64ms
    Avg time: 6082.30ms
  96th percentile:
    Connections: 1126
    Fastest time: 2828.65ms
    Slowest time: 65116.64ms
    Avg time: 6999.99ms
  97th percentile:
    Connections: 845
    Fastest time: 3196.02ms
    Slowest time: 65116.64ms
    Avg time: 8301.18ms
  98th percentile:
    Connections: 563
    Fastest time: 4122.46ms
    Slowest time: 65116.64ms
    Avg time: 10575.28ms
  99th percentile:
    Connections: 282
    Fastest time: 4603.91ms
    Slowest time: 65116.64ms
    Avg time: 16872.08ms
  99.5th percentile:
    Connections: 141
    Fastest time: 7229.20ms
    Slowest time: 65116.64ms
    Avg time: 27243.06ms
  99.9th percentile:
    Connections: 29
    Fastest time: 64545.72ms
    Slowest time: 65116.64ms
    Avg time: 64806.97ms
  99.95th percentile:
    Connections: 15
    Fastest time: 64795.59ms
    Slowest time: 65116.64ms
    Avg time: 64934.07ms
  99.99th percentile:
    Connections: 3
    Fastest time: 65030.19ms
    Slowest time: 65116.64ms
    Avg time: 65078.56ms
```

Port Pool version:
```
[Client] Connection Summary:
  Total connections attempted: 28232
  Successful connections: 28232
  Failed connections: 0
  Success rate: 100.00%
  Total runtime: 30.71 seconds
  Average connection time: 684.63ms
  Fastest connection time: 0.76ms
  Slowest connection time: 6967.61ms

[Client] Per-Decile Statistics:
  Decile 1 (10%):
    Connections: 2823
    Fastest time: 0.76ms
    Slowest time: 5.54ms
    Avg time: 3.99ms
  Decile 2 (20%):
    Connections: 2823
    Fastest time: 5.54ms
    Slowest time: 8.12ms
    Avg time: 6.77ms
  Decile 3 (30%):
    Connections: 2823
    Fastest time: 8.12ms
    Slowest time: 11.74ms
    Avg time: 9.80ms
  Decile 4 (40%):
    Connections: 2823
    Fastest time: 11.74ms
    Slowest time: 24.65ms
    Avg time: 15.55ms
  Decile 5 (50%):
    Connections: 2823
    Fastest time: 24.66ms
    Slowest time: 165.22ms
    Avg time: 104.85ms
  Decile 6 (60%):
    Connections: 2823
    Fastest time: 165.27ms
    Slowest time: 562.98ms
    Avg time: 323.67ms
  Decile 7 (70%):
    Connections: 2823
    Fastest time: 563.04ms
    Slowest time: 1090.72ms
    Avg time: 835.10ms
  Decile 8 (80%):
    Connections: 2823
    Fastest time: 1090.74ms
    Slowest time: 1560.77ms
    Avg time: 1328.20ms
  Decile 9 (90%):
    Connections: 2823
    Fastest time: 1560.80ms
    Slowest time: 1948.46ms
    Avg time: 1772.97ms
  Decile 10 (100%):
    Connections: 2825
    Fastest time: 1948.46ms
    Slowest time: 6967.61ms
    Avg time: 2444.20ms

[Client] Upper Percentile Statistics:
  90th percentile:
    Connections: 2824
    Fastest time: 1948.54ms
    Slowest time: 6967.61ms
    Avg time: 2444.37ms
  91th percentile:
    Connections: 2541
    Fastest time: 1983.29ms
    Slowest time: 6967.61ms
    Avg time: 2497.61ms
  92th percentile:
    Connections: 2259
    Fastest time: 1990.45ms
    Slowest time: 6967.61ms
    Avg time: 2561.36ms
  93th percentile:
    Connections: 1977
    Fastest time: 1995.89ms
    Slowest time: 6967.61ms
    Avg time: 2642.39ms
  94th percentile:
    Connections: 1694
    Fastest time: 2002.35ms
    Slowest time: 6967.61ms
    Avg time: 2749.85ms
  95th percentile:
    Connections: 1412
    Fastest time: 2010.97ms
    Slowest time: 6967.61ms
    Avg time: 2898.32ms
  96th percentile:
    Connections: 1130
    Fastest time: 2022.34ms
    Slowest time: 6967.61ms
    Avg time: 3118.52ms
  97th percentile:
    Connections: 847
    Fastest time: 2230.25ms
    Slowest time: 6967.61ms
    Avg time: 3475.04ms
  98th percentile:
    Connections: 565
    Fastest time: 3174.94ms
    Slowest time: 6967.61ms
    Avg time: 3854.93ms
  99th percentile:
    Connections: 283
    Fastest time: 3801.43ms
    Slowest time: 6967.61ms
    Avg time: 4160.30ms
  99.5th percentile:
    Connections: 142
    Fastest time: 3882.86ms
    Slowest time: 6967.61ms
    Avg time: 4477.64ms
  99.9th percentile:
    Connections: 29
    Fastest time: 5408.67ms
    Slowest time: 6967.61ms
    Avg time: 5682.70ms
  99.95th percentile:
    Connections: 15
    Fastest time: 5630.92ms
    Slowest time: 6967.61ms
    Avg time: 5875.13ms
  99.99th percentile:
    Connections: 3
    Fastest time: 5980.36ms
    Slowest time: 6967.61ms
    Avg time: 6619.11ms
```